backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9
aiofiles>=23.2.0
